requirements.txt
PyQt6>=6.5.0
requests>=2.28.0
beautifulsoup4>=4.12.0
python-docx>=0.8.11
Pillow>=9.0.0